lakeshore240: Move lakeshore240 agent into package
