+++ docker-compose.yml
@@ -30,15 +30,6 @@ services:
       - "socs"
 
   # --------------------------------------------------------------------------
-  # The Lakeshore 240 Agent
-  # --------------------------------------------------------------------------
-  ocs-lakeshore240-agent:
-    image: "ocs-lakeshore240-agent"
-    build: ./agents/lakeshore240/
-    depends_on:
-      - "socs"
-
-  # --------------------------------------------------------------------------
   # The Lakeshore 336 Agent
   # --------------------------------------------------------------------------
   ocs-lakeshore336-agent:
