+++ socs/agents/ocs_plugin_so.py
@@ -13,6 +13,7 @@ for n, f in [
         ('BlueforsAgent', 'bluefors/agent.py'),
         ('CryomechCPAAgent', 'cryomech_cpa/agent.py'),
         ('FTSAerotechAgent', 'fts_aerotech/agent.py'),
+        ('Lakeshore240Agent', 'lakeshore240/agent.py'),
         ('RotationAgent', 'hwp_rotation/agent.py'),
 ]:
     ocs.site_config.register_agent_class(n, os.path.join(root, f))
