+++ tests/integration/test_ls240_agent_integration.py
@@ -19,7 +19,7 @@ pytest_plugins = ("docker_compose")
 
 wait_for_crossbar = create_crossbar_fixture()
 run_agent = create_agent_runner_fixture(
-    '../agents/lakeshore240/LS240_agent.py', 'ls240_agent')
+    '../socs/agents/lakeshore240/agent.py', 'ls240_agent')
 client = create_client_fixture('LSA240S')
 
 initial_responses = {'*IDN?': 'LSCI,MODEL240,LSA240S,1.3',
